backend/requirements.txt
# Python Version Required: 3.9+

# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Firebase
firebase-admin==6.2.0

# AI & LangChain
langchain==0.1.0
langchain-openai==0.0.2
openai==1.6.1

# Machine Learning
xgboost==2.0.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2

# Technical Analysis
ta==0.11.0

# HTTP & Data APIs
yfinance>=0.2.40
httpx==0.25.2
websocket-client==1.7.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# CORS
python-cors==1.0.0
